requirements.txt
streamlit
crewai
google-genai